+++ userscript/dev.js
@@ -16,7 +16,7 @@
 // @homepageURL  https://utils.xela.codes
 // @supportURL   https://github.com/itzTheMeow/NinjaIOUtils/issues
 // @grant        none
-// @version      3.8
+// @version      3.9
 // ==/UserScript==
 
 /*