+++ userscript/src/mods/index.ts
@@ -2,6 +2,5 @@ export * from "./autoMute";
 export * from "./fpsDisplay";
 export * from "./hotkeyMessages";
 export * from "./soundEffects";
-export * from "./spectate";
 export * from "./statTracker";
 export * from "./xpStats";